requirement.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cmake==3.18.4
configparser==5.0.2
cycler==0.10.0
docker-pycreds==0.4.0
fsspec==2021.6.1
future==0.18.2
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.32.1
google-auth-oauthlib==0.4.4
grpcio==1.38.1
hydra-core==1.1.0
idna==2.10
importlib-resources==5.2.0
iniconfig==1.1.1
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
mido==1.2.10
more-itertools==8.8.0
MulticoreTSNE==0.1
multidict==5.1.0
music21==6.7.1
numpy==1.21.0
oauthlib==3.1.1
omegaconf==2.1.0
packaging==21.0
pandas==1.3.0
pathtools==0.1.2
Pillow==8.3.1
plotly==5.1.0
pluggy==0.13.1
pretty-midi==0.2.9
promise==2.3
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytorch-lightning==1.3.8
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
sentry-sdk==1.3.0
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
tenacity==8.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.2.0
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchmetrics==0.4.1
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wandb==0.10.33
webcolors==1.11.1
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
